+++ src/main/java/it/gov/pagopa/pu/fileshare/service/FileService.java
@@ -47,12 +47,6 @@ public class FileService {
 
   }
 
-  public MultipartFile getExclusivePresenceOrThrow(MultipartFile a, MultipartFile b) {
-    if (a != null && b == null) return a;
-    if (a == null && b != null) return b;
-    throw new InvalidFileException("Exactly one of the two files must be non-null");
-  }
-
   private String replaceCharVersion(String fileVersion, String target, String replacement) {
     return fileVersion.replace(target, replacement);
   }
