+++ src/main/java/it/gov/pagopa/pu/fileshare/util/Utilities.java
@@ -1,6 +1,7 @@
 package it.gov.pagopa.pu.fileshare.util;
 
 import it.gov.pagopa.pu.fileshare.dto.FileResourceDTO;
+import it.gov.pagopa.pu.fileshare.exception.custom.InvalidFileException;
 import org.slf4j.MDC;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.core.io.Resource;
@@ -8,6 +9,7 @@ import org.springframework.http.ContentDisposition;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.multipart.MultipartFile;
 
 public class Utilities {
   private Utilities(){}
@@ -29,5 +31,11 @@ public class Utilities {
 
   public static String getTraceId(){
     return MDC.get("traceId");
+  }
+
+  public static MultipartFile getExclusivePresenceOrThrow(MultipartFile a, MultipartFile b) {
+    if (a != null && b == null) return a;
+    if (a == null && b != null) return b;
+    throw new InvalidFileException("Exactly one of the two files must be non-null");
   }
 }
